Added all parts to schematic
Added all the bits and pieces needed to the shcematic and did a rough
auto route on the board. This is still assuming that the wheel only
needs 4 pins (power, ground, data1, data2). There is no circuit for
clean reading for the wheel yet. Have not run DRC or ERC.